backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
mysql-connector-python==8.0.32
python-dotenv==1.0.0
werkzeug==2.2.3  # Match with Flask 2.2.3
gunicorn==20.1.0  # For production deployment
requests==2.32.3  # For HTTP requests
boto3==1.28.15  # For AWS S3 integration